+++ src/components/ChannelsAvailabilityDialog/messages.ts
@@ -1,0 +1,10 @@
+import { defineMessages } from "react-intl";
+
+export const channelsAvailabilityDialogMessages = defineMessages({
+  noChannels: {
+    id: "mTEqYL",
+    defaultMessage:
+      "No channels to assign. Please first assign them for the product.",
+    description: "NoChannels content",
+  },
+});